backend/requirements.txt
# Веб-фреймворк
fastapi==0.104.1
uvicorn[standard]==0.24.0

# База данных и ORM
sqlalchemy==2.0.23

# Работа с данными
pydantic==1.10.13
python-multipart==0.0.6
python-dotenv==1.0.0

# Облачное хранилище
boto3==1.34.0

# Обработка изображений
pillow==10.0.0

# HTTP-запросы
requests==2.31.0